Generate website using hardcoded config file
